VSL-120: Reorganized into smaller components
